pyproject.toml
[project]
name = "llm-ensemble-classifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.12",
    "loguru>=0.7.3",
    "numpy>=2.3.0",
    "openai>=1.86.0",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.7.0",
]
